uz to simuluje vlnu dobre tak jak to ma byt proste donut s konstantnim gradientem ale stejne se to seka a jeste to neni uplne dodelane

+++ src/main/java/com/example/rp3_2d_simulace_zvuku_v_mistnosti/Pixel.java
@@ -18,11 +18,11 @@ public class Pixel extends Rectangle {
     private double realY;    // The real-world y-coordinate on the scene
     private int celkovaVychylka;
     private Rectangle rectangle;   // The visual representation of the pixel as a rectangle
-    private static double PIXELSIZE;
+    private static int PIXELSIZE;
 
     private LocalDateTime lastUpdatedTime; // To track when the pixel's color was last explicitly updated
 
-    public void setPixelSize(double pixelSize) {
+    public void setPixelSize(int pixelSize) {
         this.PIXELSIZE = pixelSize;
     }
